Fix build for OSx

+++ include/zfrog.h
@@ -533,7 +533,7 @@ struct zfrogServer
 #endif
 
 #ifndef CF_NO_HTTP
-    uint64_t    http_body_max;
+    size_t      http_body_max;
     uint16_t    http_keepalive_time;
     uint16_t    http_header_max;
     uint32_t    http_request_limit;
